# doc/api.yaml
openapi: 3.0.4
info:
  title: WASAtext
  description:  API Web Chat App, University project
  version: 1.1.1
servers:
  - url: https://api.web-chat-app.com/v1
    description: Web Chat Application API Description
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  schemas:  
...
    AuthResponse:
      type: object
      description: Standard response for Login and Registration
      properties:
        id:
          type: string
          format: uuid
          minLength: 36
          maxLength: 36
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique user identifier (UUID). Save this in the frontend state/session.
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9._~-]+$'
          description: User's username
        profilePhoto:
           type: string
           format: url
           description: URL of the profile photo (useful for UI navbar immediately after login)
      required:
        - id
        - username
...
    Error:
      type: object
      description: Textual representation of the error
      properties:
        error:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '.*'             
          description: Error message
        code:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '.*'             
          description: Error code
        details:
          type: string
          minLength: 0
          maxLength: 500
          pattern: '.*'             
          description: Additional error details
      required: 
        - error
  
## ENDPOINTS
paths:
  /session/login:
    post:
      operationId: doLogin
      security: []
      tags: 
        - Login
      summary: Logs in the user
      description: |-
        Checks if the user exists.
        - If exists: Returns the User ID and details.
        - If NOT exists: Returns 404 (User must register).
      requestBody:
        required: true
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: Schema representing the user credentials required to perform login or account creation.
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9._~-]+$'
                  example: "steph_curry"
                  description: Login Name
              required:
                - username
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                  id: "f4a8c2d1-0e9b-41b6-bd3c-527ef43ac821"
                  username: "steph_curry"
                  profilePhoto: "https://web-chat-app.com/users/profile-images/default.png"
  /session/register:                                                                                  # Registrare un nuovo utente
    post:
      operationId: registerUser
      security: []
      tags: 
        - Users
      summary: Register a new user
      description: |-
        Creates a new user account.
        Returns the user identifier and profile information upon successful registration.
        Username must be unique.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: User registration details
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9._~-]+$'
                  example: steph_curry
                  description: Unique username (letters, numbers, underscore only)
                status:
                  type: string
                  minLength: 0
                  maxLength: 200
                  pattern: '.*'
                  example: "New to WASAtext!"
                  description: Initial user status message
                profilePhoto:
                  type: string
                  format: url
                  minLength: 10
                  maxLength: 500
                  pattern: '^https?:\/\/[^\s\/$.?#].[^\s]*$'
                  description: URL to profile photo 
              required:
                - username
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                  id: "f4a8c2d1-0e9b-41b6-bd3c-527ef43ac821"
                  username: "steph_curry"
                  profilePhoto: "https://web-chat-app.com/users/profile-images/default.png"
        '400':
          description: Invalid input or username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Username already exists"
                code: "USERNAME_TAKEN"
                details: "The username 'steph_curry' is already registered"
...